CAN WORKS ON 2017 CAR
changed the FlexCAN library on the VCU to the 'Collin80' version from github to make CAN work on the teensy. Changed a || to a | to correctly handle bitshifted values in the CAN handler function o nthe VCU. Changed the message IDs on the C-CAN to reflect the correct values. Scaled the throttle input to a (0,4000) range.